Add some more features, and modularise the code slightly
There is now the possibility to change the speed the game runs at.
Currently this is hard-coded, but I plan to add a slider for this.
I also plan to add some more buttons and things to the GUI.

The mutateItem function might change, so that the ranges it
returns are within -1 and 1, but also so that the mutation is
varied enough that the neural networks actually get better.