Add an interactive UI
This brings an interactive UI with the options to pause the sim, hide individual elements or the whole thing, and change the speed at which it runs.
This also adds a system where if the only car that's left is the best one from last time, it will just kill that car.  This saves a bunch of time while training because whenever there are no improvements in a generation, you don't have to wait for the car to die to go to the next gen.