Add checkpoints
Adds checkpoints, which are used to calculate the score/fitness value (this is currently just the amount of checkpoints passed).  The fitness value is used to choose the car to base the next generation on.
Also adds the functionality of automatically killing the best car, so training is faster.  This is toggleable.  There is also a manual kill button, which just kills every car, so as to start a new generation.
Also also adds a visual indicator of the mouses current coordinates, so that making the track and checkpoints is easier.