Reorganise UI into tables
Reorganises the UI into tables.
Also adds a value, time, to the values that are show to the user.  This shows the time in this generation, from the cars' point of view.  In other words, if the speed is turned up, the time will count up faster.
Adds a training mode, which just sets some of the user-controllable controls to ones that make training the fastest.